Move directory structure
This now matches the AWS and K8s sources
