+++ adapters/ip_cache_test.go
@@ -1,4 +1,4 @@
-package internet
+package adapters
 
 import (
 	"net"
