+++ adapters/rdap-asn.go
@@ -1,4 +1,4 @@
-package internet
+package adapters
 
 import (
 	"context"
@@ -10,22 +10,22 @@ import (
 	"github.com/overmindtech/sdpcache"
 )
 
-type ASNAdapter struct {
+type RdapASNAdapter struct {
 	ClientFac func() *rdap.Client
 	Cache     *sdpcache.Cache
 }
 
 // Type is the type of items that this returns
-func (s *ASNAdapter) Type() string {
+func (s *RdapASNAdapter) Type() string {
 	return "rdap-asn"
 }
 
 // Name Returns the name of the backend
-func (s *ASNAdapter) Name() string {
+func (s *RdapASNAdapter) Name() string {
 	return "rdap"
 }
 
-func (s *ASNAdapter) Metadata() *sdp.AdapterMetadata {
+func (s *RdapASNAdapter) Metadata() *sdp.AdapterMetadata {
 	adapter := AutumnMetadata()
 	return &adapter
 }
@@ -44,17 +44,17 @@ func AutumnMetadata() sdp.AdapterMetadata {
 }
 
 // Weighting of duplicate adapters
-func (s *ASNAdapter) Weight() int {
+func (s *RdapASNAdapter) Weight() int {
 	return 100
 }
 
-func (s *ASNAdapter) Scopes() []string {
+func (s *RdapASNAdapter) Scopes() []string {
 	return []string{
 		"global",
 	}
 }
 
-func (s *ASNAdapter) Get(ctx context.Context, scope string, query string, ignoreCache bool) (*sdp.Item, error) {
+func (s *RdapASNAdapter) Get(ctx context.Context, scope string, query string, ignoreCache bool) (*sdp.Item, error) {
 	hit, ck, items, sdpErr := s.Cache.Lookup(ctx, s.Name(), sdp.QueryMethod_GET, scope, s.Type(), query, ignoreCache)
 
 	if sdpErr != nil {
@@ -81,7 +81,7 @@ func (s *ASNAdapter) Get(ctx context.Context, scope string, query string, ignore
 	if err != nil {
 		err = wrapRdapError(err)
 
-		s.Cache.StoreError(err, CacheDuration, ck)
+		s.Cache.StoreError(err, RdapCacheDuration, ck)
 
 		return nil, err
 	}
@@ -133,12 +133,12 @@ func (s *ASNAdapter) Get(ctx context.Context, scope string, query string, ignore
 
 	item.LinkedItemQueries = extractEntityLinks(asn.Entities)
 
-	s.Cache.StoreItem(item, CacheDuration, ck)
+	s.Cache.StoreItem(item, RdapCacheDuration, ck)
 
 	return item, nil
 }
 
-func (s *ASNAdapter) List(ctx context.Context, scope string, ignoreCache bool) ([]*sdp.Item, error) {
+func (s *RdapASNAdapter) List(ctx context.Context, scope string, ignoreCache bool) ([]*sdp.Item, error) {
 	return nil, &sdp.QueryError{
 		ErrorType:   sdp.QueryError_NOTFOUND,
 		Scope:       scope,
