+++ adapters/rdap-domain_test.go
@@ -1,4 +1,4 @@
-package internet
+package adapters
 
 import (
 	"context"
@@ -11,7 +11,7 @@ import (
 func TestDomainAdapterGet(t *testing.T) {
 	t.Parallel()
 
-	src := &DomainAdapter{
+	src := &RdapDomainAdapter{
 		ClientFac: func() *rdap.Client { return testRdapClient(t) },
 		Cache:     sdpcache.NewCache(),
 	}
