+++ adapters/rdap-nameserver.go
@@ -1,4 +1,4 @@
-package internet
+package adapters
 
 import (
 	"context"
@@ -10,27 +10,27 @@ import (
 	"github.com/overmindtech/sdpcache"
 )
 
-type NameserverAdapter struct {
+type RdapNameserverAdapter struct {
 	ClientFac func() *rdap.Client
 	Cache     *sdpcache.Cache
 }
 
 // Type is the type of items that this returns
-func (s *NameserverAdapter) Type() string {
+func (s *RdapNameserverAdapter) Type() string {
 	return "rdap-nameserver"
 }
 
 // Name Returns the name of the adapter
-func (s *NameserverAdapter) Name() string {
+func (s *RdapNameserverAdapter) Name() string {
 	return "rdap"
 }
 
 // Weighting of duplicate adapters
-func (s *NameserverAdapter) Weight() int {
+func (s *RdapNameserverAdapter) Weight() int {
 	return 100
 }
 
-func (s *NameserverAdapter) Metadata() *sdp.AdapterMetadata {
+func (s *RdapNameserverAdapter) Metadata() *sdp.AdapterMetadata {
 	adapter := NameserverMetadata()
 	return &adapter
 }
@@ -47,13 +47,13 @@ func NameserverMetadata() sdp.AdapterMetadata {
 		Category:       sdp.AdapterCategory_ADAPTER_CATEGORY_NETWORK,
 	}
 }
-func (s *NameserverAdapter) Scopes() []string {
+func (s *RdapNameserverAdapter) Scopes() []string {
 	return []string{
 		"global",
 	}
 }
 
-func (s *NameserverAdapter) Get(ctx context.Context, scope string, query string, ignoreCache bool) (*sdp.Item, error) {
+func (s *RdapNameserverAdapter) Get(ctx context.Context, scope string, query string, ignoreCache bool) (*sdp.Item, error) {
 	// Check the cache for GET requests, if we don't hit the cache then there is
 	// nothing we can do though
 	hit, _, items, sdpErr := s.Cache.Lookup(ctx, s.Name(), sdp.QueryMethod_GET, scope, s.Type(), query, ignoreCache)
@@ -77,7 +77,7 @@ func (s *NameserverAdapter) Get(ctx context.Context, scope string, query string,
 	}
 }
 
-func (s *NameserverAdapter) List(ctx context.Context, scope string, ignoreCache bool) ([]*sdp.Item, error) {
+func (s *RdapNameserverAdapter) List(ctx context.Context, scope string, ignoreCache bool) ([]*sdp.Item, error) {
 	return nil, &sdp.QueryError{
 		ErrorType:   sdp.QueryError_NOTFOUND,
 		Scope:       scope,
@@ -90,7 +90,7 @@ func (s *NameserverAdapter) List(ctx context.Context, scope string, ignoreCache 
 // which nameserver to query from the beginning. Fortunately domain queries can
 // be bootstrapped, so we can use the domain query to find the nameserver in the
 // link
-func (s *NameserverAdapter) Search(ctx context.Context, scope string, query string, ignoreCache bool) ([]*sdp.Item, error) {
+func (s *RdapNameserverAdapter) Search(ctx context.Context, scope string, query string, ignoreCache bool) ([]*sdp.Item, error) {
 	hit, ck, items, sdpErr := s.Cache.Lookup(ctx, s.Name(), sdp.QueryMethod_SEARCH, scope, s.Type(), query, ignoreCache)
 
 	if sdpErr != nil {
@@ -118,7 +118,7 @@ func (s *NameserverAdapter) Search(ctx context.Context, scope string, query stri
 	if err != nil {
 		err = wrapRdapError(err)
 
-		s.Cache.StoreError(err, CacheDuration, ck)
+		s.Cache.StoreError(err, RdapCacheDuration, ck)
 
 		return nil, err
 	}
@@ -203,7 +203,7 @@ func (s *NameserverAdapter) Search(ctx context.Context, scope string, query stri
 		}
 	}
 
-	s.Cache.StoreItem(item, CacheDuration, ck)
+	s.Cache.StoreItem(item, RdapCacheDuration, ck)
 
 	return []*sdp.Item{item}, nil
 }
